backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-decouple==3.8
pymysql==1.1.0
sqlalchemy==2.0.23
python-jose==3.3.0
python-multipart==0.0.6
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0